% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf function.R
\name{accept.reject}
\alias{accept.reject}
\title{Martino Prob Final}
\usage{
accept.reject(n, pdf, a, b, C)
}
\arguments{
\item{n}{is the desired number of samples}

\item{pdf}{is the pdf of the random variable that you wish to sample from}

\item{a}{is the lower bound of the random variable}

\item{b}{is the upper bound of the random variable}

\item{C}{is the number that f(x) <= C for all x}
}
\value{
the samples
}
\description{
samples a give pdf of a random variable
}
